[Team-03][Android]  네트워크 코드 수정

+++ android/app/src/main/java/com/example/todo/respository/ToDoRepository.kt
@@ -3,10 +3,9 @@ package com.example.todo.respository
 import android.util.Log
 import com.example.todo.model.ProgressType
 import com.example.todo.model.TodoItem
-import com.example.todo.network.Done
-import com.example.todo.network.Ing
 import com.example.todo.network.Todo
 import com.example.todo.network.TodoResponseItem
+import retrofit2.Response
 
 class ToDoRepository(private val toDoDataSource: ToDoDataSource) {
 
@@ -31,7 +30,7 @@ class ToDoRepository(private val toDoDataSource: ToDoDataSource) {
         return list.map { TodoItem(it.title, it.content, ProgressType.TO_DO, it.id, it.nextId) }
     }
 
-    private fun jsonInProgressListToTodoItems(list: List<Ing>): List<TodoItem> {
+    private fun jsonInProgressListToTodoItems(list: List<Todo>): List<TodoItem> {
         return list.map {
             TodoItem(
                 it.title,
@@ -43,25 +42,34 @@ class ToDoRepository(private val toDoDataSource: ToDoDataSource) {
         }
     }
 
-    private fun jsonDoneToTodoItems(list: List<Done>): List<TodoItem> {
+    private fun jsonDoneToTodoItems(list: List<Todo>): List<TodoItem> {
         return list.map { TodoItem(it.title, it.content, ProgressType.DONE, it.id, it.nextId) }
     }
 
     suspend fun addToDoItem(toDoList: List<TodoItem>, newItem: TodoItem): List<TodoItem> {
         val response = toDoDataSource.getTodoId(newItem)
+        Log.d("testApi", response.isSuccessful.toString())
         return if (response.isSuccessful) {
             val originList = toDoList.toMutableList()
-            newItem.itemId = response?.body()?.card_id ?: -1
-            originList[1].next = newItem.itemId
+
+            newItem.itemId = response?.body()?.cardId ?: -1
+            originList[originList.size - 1].next = newItem.itemId
+
             originList.add(0, newItem)
             originList.toList()
         } else toDoList
     }
 
-    fun deleteToDoItem(toDoList: List<TodoItem>, deleteItem: TodoItem): List<TodoItem> {
+    suspend fun removeToDoItem(toDoList: List<TodoItem>, deleteItem: TodoItem): List<TodoItem> {
         val originList = toDoList.toMutableList()
-        originList.remove(deleteItem)
-        return originList.toList()
+        deleteItem.itemId?.let {
+            val response = toDoDataSource.removeItem(it.toInt())
+            if (response.isSuccessful) {
+                Log.d("testDelete", "ddd")
+                originList.remove(deleteItem)
+            }
+            return originList.toList()
+        } ?: kotlin.run { return originList.toList() }
     }
 
     fun updateToDoItem(toDoList: List<TodoItem>, updateItem: TodoItem): List<TodoItem> {
